correct a bad search and replace

+++ app/controllers/works_controller.rb
@@ -4,102 +4,102 @@ class WorksController < ApplicationController
   before_action :load_assignment
   
   def load_semester
-   xxx@xxx.xxx = Semester.find( params[:semester_id] )
+    @semester = Semester.find( params[:semester_id] )
   end
   
   def load_section
-   xxx@xxx.xxx = Section.find( params[:section_id] )
+    @section = Section.find( params[:section_id] )
   end
 
   def load_assignment
-   xxx@xxx.xxx = Assignment.find( params[:assignment_id] )
+    @assignment = Assignment.find( params[:assignment_id] )
   end  
   
   # GET /works/new
   # GET /works/new.xml
   def new
-    @work xxx@xxx.xxx
+    @work = @assignment.works.build
 
     respond_to do |format|
       format.html # new.html.erb
-      format.xml  { render xxx@xxx.xxx }
+      format.xml  { render xml: @work }
     end
   end
 
   # POST /works
   # POST /works.xml
   def create
-    @work xxx@xxx.xxx
-   xxx@xxx.xxx
+    @work = @assignment.works.build(allowed_params)
+    @work.email.downcase!
 
     respond_to do |format|
       # make this work one with the assignment
-      @work.assignment_id xxx@xxx.xxx
+      @work.assignment_id = @assignment.id
       
       # try to find the student who this assignment belongs to (based on the submitting email address)
       student = Student.where('email=?', params[:work][:email].downcase).first
       
       if student
         # if a student is found with a matching email address attribute the assignment to them
-        @work.enrollment = xxx@xxx.xxx
+        @work.enrollment = student.enrollments.where('section_id=?',@section.id).first
       end
       
       # check if a work has already been submitted from this submission address
-      prev_works xxx@xxx.xxx params[:work][:email].downcase)
+      prev_works = @assignment.works.where('email=?', params[:work][:email].downcase)
 
       # if the work wasn't previously submitted
       if prev_works.size == 0
         # save the work
-        xxx@xxx.xxx
+        if @work.save
           flash[:notice] = 'Your work was successfully uploaded.'
         
           # send the student an email including the withdrawal link
-         xxx@xxx.xxx params[:work][:email].downcase
+          @work.send_submitted_email params[:work][:email].downcase
         
-          @status = xxx@xxx.xxx
-          format.html { xxx@xxx.xxx } 
-          format.xml  { render xml: @work, status: :created, xxx@xxx.xxx }
+          @status = semester_section_assignment_url(@semester,@section,@assignment)
+          format.html { redirect_to([@semester,@section,@assignment]) } 
+          format.xml  { render xml: @work, status: :created, location: @work }
         else
-         xxx@xxx.xxx = "failed"
+          @status = "failed"
           format.html { render action: "new" }
-          format.xml  { render xxx@xxx.xxx status: :unprocessable_entity }
+          format.xml  { render xml: @work.errors, status: :unprocessable_entity }
         end
       else
-         xxx@xxx.xxx :email, 'has already submitted this work. If you want to resubmit, you must first withdraw your previous submission (check your email)'
+          @work.errors.add :email, 'has already submitted this work. If you want to resubmit, you must first withdraw your previous submission (check your email)'
         
           # send the submitter a reminder including the withdrawal link
           prev_works[0].send_reminder_email # params[:work][:email]
         
-         xxx@xxx.xxx = "failed"
+          @status = "failed"
           format.html { render action: "new" }
-          format.xml  { render xxx@xxx.xxx status: :unprocessable_entity }
+          format.xml  { render xml: @work.errors, status: :unprocessable_entity }
       end
       format.js
     end
   end
   
   def withdraw
-   xxx@xxx.xxx = ""
+    @condition = ""
     
     if params[:code]
-      @works xxx@xxx.xxx
+      @works = @assignment.works.where('withdrawal_code=?',params[:code])
       
-      xxx@xxx.xxx > 0
-       xxx@xxx.xxx do |work|
+      if @works.size > 0
+        @works.each do |work|
           if work.withdrawal_code == params[:code]
             if work.assignment.duedate < Time.now
-             xxx@xxx.xxx = "pastdue"
+              @condition = "pastdue"
             else
               work.destroy
-             xxx@xxx.xxx = "destroyed"
+              @condition = "destroyed"
             end
           end
         end
       else
-       xxx@xxx.xxx = "no works"  
+        @condition = "no works"  
       end
     else
-     xxx@xxx.xxx = "no code"
+      @condition = "no code"
     end
     
     respond_to do |format|
